requirements.txt
flask==2.1.2
flask_cors==3.0.10
pymongo==4.1.1
PyYAML==5.4.1